Adjust rendering to limit reverse dependency depth.
